+++ tool/build.dart
@@ -1,0 +1,6 @@
+import 'dart:io';
+
+void main() {
+  var uri = Platform.script;
+  print(uri.toFilePath());
+}