+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 0.0.1
 
-- Initial version, created by Stagehand
+- Initial version
+- Based on libsass v3.1.0
+- Wrapped libsass functions in Dart classes